v3.1.2
- Separate behaviors to other files
