+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.1.1",
+  "version": "3.1.2",
   "name": "gcc-min",
   "description": "Just write scripts for Node.js environment, then build to use in the browsers",
   "repository": {
@@ -28,7 +28,7 @@
     "babel-plugin-transform-remove-strict-mode": "0.x.x",
     "babel-preset-env": "1.x.x",
     "bellajs": "6.x.x",
-    "butternut": "0.x.x"
+    "uglify-js": "3.x.x"
   },
   "devDependencies": {
     "coveralls": "2.x.x",
