+++ src/minify.js
@@ -1,0 +1,11 @@
+/**
+ * minify
+ * @ndaidong
+**/
+
+var {minify} = require('uglify-js');
+
+module.exports = (source = '') => {
+  let {code} = minify(source);
+  return code;
+};
